+++ src/components/ThemeSwitch.jsx
@@ -1,46 +1,56 @@
 import { useEffect, useState } from "react";
 import { Switch } from "@headlessui/react";
 import { SunIcon } from "@heroicons/react/24/solid";
-import { useLocalStorage } from "usehooks-ts";
 
+//i dont know about this, it's a ts stuff
 function classNames(...classes) {
   return classes.filter(Boolean).join(" ");
 }
 
 function ThemeSwitch() {
-  const [theme, setTheme] = useLocalStorage("theme", "light");
+  const [darkMode, setDarkMode] = useState(localStorage.theme !== "dark");
 
   useEffect(() => {
-    document.body.classList.remove("light", "dark");
-    document.body.classList.add(theme);
-  }, [theme]);
+    if (localStorage.theme === "dark") {
+      document.documentElement.classList.add("dark");
+      console.log("added dark class");
+    } else {
+      //document.documentElement.classList.remove("dark");
+      console.log("no dark class in local storage");
+    }
+  }, [darkMode]);
 
-  const [enabled, setEnabled] = useState(theme == "light");
-
-  const handleThemeChange = (enabled) => {
-    setTheme(enabled ? "light" : "dark");
-    setEnabled(enabled);
+  //change theme
+  const toggleThemeChange = () => {
+    setDarkMode(!darkMode);
+    if (darkMode) {
+      localStorage.setItem("theme", "dark");
+      document.documentElement.classList.add("dark");
+    } else {
+      localStorage.setItem("theme", null);
+      document.documentElement.classList.remove("dark");
+    }
   };
 
   return (
     <Switch
-      checked={enabled}
-      onChange={handleThemeChange}
+      checked={darkMode}
+      onChange={toggleThemeChange}
       className={classNames(
-        enabled ? "bg-gray-400" : "bg-yellow-600",
+        !darkMode ? "bg-gray-400" : "bg-yellow-600",
         "relative inline-flex h-6 w-11 flex-shrink-0 cursor-pointer rounded-full border-2 border-transparent transition-colors duration-200 ease-in-out"
       )}
     >
       <span className="sr-only">Use setting</span>
       <span
         className={classNames(
-          enabled ? "translate-x-5" : "translate-x-0",
+          darkMode ? "translate-x-5" : "translate-x-0",
           "pointer-events-none relative inline-block h-5 w-5 transform rounded-full bg-white shadow ring-0 transition duration-200 ease-in-out"
         )}
       >
         <span
           className={classNames(
-            enabled
+            darkMode
               ? "opacity-0 duration-100 ease-out"
               : "opacity-100 duration-200 ease-in",
             "absolute inset-0 flex h-full w-full items-center justify-center transition-opacity"
@@ -51,7 +61,7 @@ function ThemeSwitch() {
         </span>
         <span
           className={classNames(
-            enabled
+            darkMode
               ? "opacity-100 duration-200 ease-in"
               : "opacity-0 duration-100 ease-out",
             "absolute inset-0 flex h-full w-full items-center justify-center transition-opacity"
